requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
pyzbar>=0.1.9
RPi.GPIO>=0.7.1
numpy>=1.24.0
Pillow>=10.0.0

